# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HifzMaal API",
        "description": "Islamic Family Finance Management Platform",
        "contact": {
            "email": "support@hifzmaal.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        },
        {
            "url": "https://api.hifzmaal.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/families/{family}/bills": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get all bills for a family",
                "operationId": "394181b6beb51104aa2bfc1240ee84d8",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "paid",
                                "overdue"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "electricity",
                                "gas",
                                "water",
                                "internet",
                                "mobile",
                                "rent",
                                "school_fees",
                                "other"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bills"
                ],
                "summary": "Create a new bill",
                "operationId": "61aec20eb5b92e465210e42147d9788e",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "type",
                                    "amount",
                                    "due_date",
                                    "frequency"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Electricity Bill"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "electricity",
                                            "gas",
                                            "water",
                                            "internet",
                                            "mobile",
                                            "rent",
                                            "school_fees",
                                            "other"
                                        ],
                                        "example": "electricity"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3500
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-02-15"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "monthly",
                                            "quarterly",
                                            "yearly"
                                        ],
                                        "example": "monthly"
                                    },
                                    "is_recurring": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "auto_pay": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "account_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "provider": {
                                        "type": "string",
                                        "example": "K-Electric"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "split_members": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "reminder_days": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bill created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/families/{family}/bills/{bill}": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get bill by ID",
                "operationId": "90a9effc242d8398cdd0a2df92610b04",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bill",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bill not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bills"
                ],
                "summary": "Update bill",
                "operationId": "65f53e081378285c4fcff90a6e7e051f",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bill",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "reminder_days": {
                                        "type": "integer"
                                    },
                                    "auto_pay": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bill updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bills"
                ],
                "summary": "Delete bill",
                "operationId": "ab12c21be45908231217676be8fae97e",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bill",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bill deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/families/{family}/bills/{bill}/mark-as-paid": {
            "post": {
                "tags": [
                    "Bills"
                ],
                "summary": "Mark bill as paid",
                "operationId": "3ef3a487cb8c6f9132bc0d6a84b71e1c",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bill",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bill marked as paid successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/families/{family}/bills/upcoming": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get upcoming bills",
                "operationId": "4363b3b79402cf95b42b33047ec41d6e",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/families/{family}/bills/overdue": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get overdue bills",
                "operationId": "7c2784b3881bb0b72e58d1b6c36bca44",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/families/{family}/bills/statistics": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get bill statistics",
                "operationId": "d0b397d339bdaf4a9714febdcb9d6763",
                "parameters": [
                    {
                        "name": "family",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_bills": {
                                                    "type": "integer"
                                                },
                                                "pending_bills": {
                                                    "type": "integer"
                                                },
                                                "overdue_bills": {
                                                    "type": "integer"
                                                },
                                                "paid_this_month": {
                                                    "type": "integer"
                                                },
                                                "total_due_amount": {
                                                    "type": "number"
                                                },
                                                "average_bill_amount": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}